Grant our bot "push" (write) permissions on setup-cloud-sdk
The entire context is at https://github.com/orgs/community/discussions/57972#discussioncomment-6201394, but GitHub published a breaking change without notice, and we have to grant out Bot overzealous permissions as a result. This is the only repo affected.

I also found a small bug in our merge logic, so that's fixed. This is already applied.